net: emaclite: Use unsigned long for baseaddr
Baseaddr should be unsigned long.

+++ include/netdev.h
@@ -90,7 +90,8 @@ int smc91111_initialize(u8 dev_num, int base_addr);
 int tsi108_eth_initialize(bd_t *bis);
 int uec_standard_init(bd_t *bis);
 int uli526x_initialize(bd_t *bis);
-int xilinx_emaclite_initialize (bd_t *bis, int base_addr, int txpp, int rxpp);
+int xilinx_emaclite_initialize(bd_t *bis, unsigned long base_addr,
+							int txpp, int rxpp);
 int xilinx_ll_temac_initialize(bd_t *bis, int base_addr);
 int xilinx_axiemac_initialize(bd_t *bis, unsigned long base_addr, int dma_addr);
 int sh_eth_initialize(bd_t *bis);
